+++ packages/server/src/middleware/__tests__/funcs/fallback404.func.ts
@@ -1,0 +1,8 @@
+import { staticHandler, useMiddleware } from '../../../middleware'
+
+export const func = useMiddleware(
+  staticHandler({
+    root: __dirname,
+    notFound: 'useMiddleware.func.ts',
+  })
+)