+++ packages/server/src/middleware/__tests__/index.test.ts
@@ -1,4 +1,3 @@
-import { readFileSync } from 'node:fs'
 import { join } from 'node:path'
 import { request } from '@faasjs/request'
 import { afterAll, beforeAll, describe, expect, it } from 'vitest'
@@ -64,75 +63,6 @@ describe('middleware', () => {
       statusCode: 500,
       headers: {},
       body: 'Error: useMiddleware',
-    })
-  })
-
-  describe('staticHandler', () => {
-    it('should work', async () => {
-      await expect(
-        request(`http://127.0.0.1:${port}/useMiddleware.func.ts`)
-      ).resolves.toMatchObject({
-        statusCode: 200,
-        headers: {},
-        body: readFileSync(
-          join(__dirname, 'funcs', 'useMiddleware.func.ts'),
-          'utf-8'
-        ),
-      })
-      await expect(
-        request(`http://127.0.0.1:${port}/useMiddleware.func.ts`)
-      ).resolves.toMatchObject({
-        statusCode: 200,
-        headers: {},
-        body: readFileSync(
-          join(__dirname, 'funcs', 'useMiddleware.func.ts'),
-          'utf-8'
-        ),
-      })
-      await expect(
-        request(`http://127.0.0.1:${port}/default.default.func.ts`)
-      ).resolves.toMatchObject({
-        statusCode: 200,
-        headers: {},
-        body: readFileSync(
-          join(__dirname, 'funcs', 'default.func.ts'),
-          'utf-8'
-        ),
-      })
-    })
-
-    it('not found', async () => {
-      await expect(
-        request(`http://127.0.0.1:${port}/404`)
-      ).rejects.toMatchObject({
-        statusCode: 404,
-        headers: {},
-        body: 'Not Found',
-      })
-      await expect(
-        request(`http://127.0.0.1:${port}/404`)
-      ).rejects.toMatchObject({
-        statusCode: 404,
-        headers: {},
-        body: 'Not Found',
-      })
-    })
-
-    it('custom not found', async () => {
-      await expect(
-        request(`http://127.0.0.1:${port}/custom404`)
-      ).rejects.toMatchObject({
-        statusCode: 404,
-        headers: {},
-        body: 'custom404',
-      })
-      await expect(
-        request(`http://127.0.0.1:${port}/custom404`)
-      ).rejects.toMatchObject({
-        statusCode: 404,
-        headers: {},
-        body: 'custom404',
-      })
     })
   })
 })